Update and rename HelloController.java to MainController.java
I created our main controller class. This class includes treeView, search and filter search parts on the left side. For the search part, I used both BST and partial searching. On the right side, we have a screen for whatever is selected on treeView. For Type and Tag objects, we have their name, a button to change their name, and a delete button. For the Item objects, additionally, we have a type name label, tag name label, table view for properties of the item, and an edit button that opens a new screen to edit selected item's variables. In our menu bar, we have Create menu. When the user clicks it, create scene opens for the user to create an item.